+++ CHANGELOG.md
@@ -9,10 +9,12 @@ This is a changelog for **ZachranObed** application.
 ### Fixed
 - **ZOB-72** Add scroll padding for password field.
 - **ZOB-113** Lock Android for portrait screen orientation in Manifest.
+- **ZOB-219** Add validation for consume-by date.
 
 ### Changed
 - **ZOB-296** Update allergens UI in food offer form.
 - **ZOB-298** Update food categories and box types UI in food offer form.
+- **ZOB-300** Update consume-by UI in food offer form.
 
 ### Removed
 - **ZOB-295** Remove updating delivery state to NOT_USED from application.